added product edition logic

+++ excluir-produto.php
@@ -4,10 +4,12 @@ require 'vendor/autoload.php';
 
 use Diogopachecodev\SerenattoCoffee\Infrastructure\Repository\PdoProductRepository;
 
-$pdo = Diogopachecodev\SerenattoCoffee\Infrastructure\Persistence\Database::connect();
-$productRepository = new PdoProductRepository($pdo);
+if(!empty($_POST)) {
+    $pdo = Diogopachecodev\SerenattoCoffee\Infrastructure\Persistence\Database::connect();
+    $productRepository = new PdoProductRepository($pdo);
 
-$productRepository->deleteProduct($_POST['id']);
+    $productRepository->deleteProduct($_POST['id']);
+}
 
 header('Location: ./admin.php');
 exit();
